project/requirements.txt
grpcio==1.34.0
grpcio-tools==1.34.0
protobuf==3.14.0
six==1.15.0
